GLoft/requirements.txt
asgiref==3.3.1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
coveralls==3.0.1
cryptography==3.2.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
docopt==0.6.2
drf-yasg==1.20.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
static3==0.7.0
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
